msm: mdss: Allow mdss driver to carry more lazy buffers
SMMU Virtulization limit depends on the page table format
for a 32bit table we can have 4GB as a limit.

Increase the virtual map range for mdss driver so that
it can allow more lazy buffers to be mapped.

This will help in cases where apps which are queued
once to display and are parked in background.

Change-Id: I08f8cbff267630f0ade33299762c4900c04f611b
Signed-off-by: Raviteja Tamatam <travitej@codeaurora.org>
Signed-off-by: engstk <eng.stk@sapo.pt>